# api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Where did my money go
servers:
  - url: /api/v1
paths:
  /movement:
    post:
      summary: Create a movement
      operationId: createMovement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movement'
        required: true
      responses:
        '200':
          description: Ok
    get:
      summary: Get all movements
      operationId: getAllMovement
      responses:
        '200':
          description: get movements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListMovements"
    put:
      summary: Update a movement
      operationId: updateMovement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movement'
        required: true
      responses:
        '200':
          description: Movement updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movement"
        '404':
          description: Movement not found
  /movement/{year}/{month}:
    get:
      summary: Get all the movements by year and month
      operationId: getMovementsByYearAndMonth
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
          example: 2025
        - name: month
          in: path
          required: true
          schema:
            type: string
          example: 5
      responses:
        '200':
          description: get movements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListMovements"
  /movement/{id}:
    delete:
      summary: Delete a movement
      operationId: deleteMovement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Movement deleted

components:
  schemas:
    Movement:
      type: object
      required:
        - quantity
        - reason
        - subreason
        - necessary
        - dateMovement
        - id
      properties:
        reason:
          $ref: "#/components/schemas/TypeMovement"
        necessary:
          type: boolean
        quantity:
          type: double
        subreason:
          type: string
        dateCreation:
          type: dateTime
        dateMovement:
          type: dateTime
        id:
          type: long
    TypeMovement:
      enum:
        - TRABAJO
        - OCIO
        - SALUD
        - CAPRICHO
        - INVERSION
        - DEPORTE
        - COMIDA
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    ListMovements:
      type: array
      items:
        $ref: "#/components/schemas/Movement"